update json operations by creating own class for it
from now app is not using any external library as now it has it's own json reading and writing system for that
app had did this to overcome some of the build issues it is facing
however now app can't read an list from json file but can write list to json. We can extent json reading behavior if required in future.