fix previous log panel not updating when new window opens
the issue I fix was that when we open a new window then the log panel of the previous window does not update it self this is because it was the responsibility of the main class i.e. app class to add logs to it. But now this responsibility is given to the log panel itself.